# TestConsoleApp/OpenAPIs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Api服务在线说明"
  },
  "host": "192.168.2.35:8050",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/MesService/MaterialReqOut": {
      "post": {
        "tags": [
          "MesService"
        ],
        "summary": "MES请求原材料出库-自动出库",
        "operationId": "MesService_MaterialReqOut",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MesRequestOutboundTaskBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/MesService/MaterialReqManuallyOut": {
      "post": {
        "tags": [
          "MesService"
        ],
        "summary": "MES请求原材料出库-人工出库",
        "operationId": "MesService_MaterialReqManuallyOut",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MesRequestManuallyOutboundBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/MesService/IssuePlanQuery": {
      "post": {
        "tags": [
          "MesService"
        ],
        "summary": "MES请求原材料出库计划查询",
        "operationId": "MesService_IssuePlanQuery",
        "consumes": [],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "planCode",
            "in": "query",
            "description": "要查询的计划号",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MesIssuePlanQuery"
            }
          }
        }
      }
    },
    "/api/MesService/MaterialReqReturn": {
      "post": {
        "tags": [
          "MesService"
        ],
        "summary": "MES请求原材料返库",
        "operationId": "MesService_MaterialReqReturn",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MesRequestReturnTaskBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/MesService/StationRotation": {
      "post": {
        "tags": [
          "MesService"
        ],
        "summary": "MES发送机台旋转通知",
        "operationId": "MesService_StationRotation",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MesStationRotationReportBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/MesService/ChangeInventory": {
      "post": {
        "tags": [
          "MesService"
        ],
        "summary": "MES发送机台旋转通知",
        "operationId": "MesService_ChangeInventory",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MesRequestChangeInventoryBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/MesService/FinishedProductReqIn": {
      "post": {
        "tags": [
          "MesService"
        ],
        "summary": "MES请求成品完工入库",
        "operationId": "MesService_FinishedProductReqIn",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductInboundRequestBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "WmsResponseBody": {
      "description": "wms服务接口返回信息",
      "type": "object",
      "properties": {
        "TaskCode": {
          "description": "请求任务号",
          "type": "string"
        },
        "Code": {
          "description": "状态码",
          "type": "string"
        },
        "Msg": {
          "description": "消息",
          "type": "string"
        }
      }
    },
...
    "MesRequestOutboundTaskBody": {
      "description": "MES请求物料自动出库任务信息实体类",
      "type": "object",
      "properties": {
        "TaskCode": {
          "description": "申请任务号",
          "type": "string"
        },
        "StationCode": {
          "description": "申请机台号",
          "type": "string"
        },
        "ReqTime": {
          "description": "申请时间",
          "type": "string"
        },
        "TaskList": {
          "description": "任务明细信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MesRequestTaskDetailBody"
          }
        }
      }
    },
    "MesRequestTaskDetailBody": {
      "description": "Mes请求出库任务明细信息实体类",
      "type": "object",
      "properties": {
        "LineCode": {
          "description": "机台关联的输送线编号",
          "type": "string"
        },
        "MaterialList": {
          "description": "物料明细信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MesRequestMaterialBody"
          }
        }
      }
    },
    "MesRequestMaterialBody": {
      "description": "物料信息实体类",
      "type": "object",
      "properties": {
        "MaterialCode": {
          "description": "物料编码",
          "type": "string"
        },
        "MaterialName": {
          "description": "物料名称",
          "type": "string"
        },
        "MaterialBarcode": {
          "description": "物料条码",
          "type": "string"
        },
        "Qty": {
          "format": "double",
          "description": "出库申请：申请数量；返库申请：WMS送料时传入的数量",
          "type": "number"
        },
        "ConsumedQty": {
          "format": "double",
          "description": "物料消耗数量",
          "type": "number"
        },
        "EmptyFlag": {
          "description": "空轴标识（导线和端子用）；Y: 空   N: 非空；如果是模具，则默认填写'N'。",
          "type": "string"
        },
        "MaterialPS": {
          "description": "包装规格；成品入库时用",
          "type": "string"
        }
      }
    },
    "MesRequestManuallyOutboundBody": {
      "description": "MES请求物料人工出库任务信息实体类",
      "type": "object",
      "properties": {
        "TaskCode": {
          "description": "申请任务号",
          "type": "string"
        },
        "StationCode": {
          "description": "申请机台号",
          "type": "string"
        },
        "ReqTime": {
          "description": "申请时间",
          "type": "string"
        },
        "MaterialList": {
          "description": "任务明细信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ManuallyOutboundMaterialBody"
          }
        }
      }
    },
    "ManuallyOutboundMaterialBody": {
      "description": "物料信息",
      "type": "object",
      "properties": {
        "MaterialCode": {
          "description": "物料编码",
          "type": "string"
        },
        "Qty": {
          "format": "double",
          "description": "出库申请：申请数量；返库申请：WMS送料时传入的数量",
          "type": "number"
        }
      }
    },
    "MesIssuePlanQuery": {
      "description": "领料计划查询",
      "type": "object",
      "properties": {
        "PlanCode": {
          "description": "申请计划号",
          "type": "string"
        },
        "Code": {
          "description": "状态码",
          "type": "string"
        },
        "Msg": {
          "description": "消息",
          "type": "string"
        },
        "PlanList": {
          "description": "计划明细",
          "type": "array",
          "items": {
            "$ref": "#/definitions/IssuePlanList"
          }
        }
      }
    },
    "IssuePlanList": {
      "description": "计划明细信息",
      "type": "object",
      "properties": {
        "MaterialCode": {
          "description": "物料编码",
          "type": "string"
        },
        "IssueQty": {
          "format": "double",
          "description": "发料数量",
          "type": "number"
        }
      }
    },
    "MesRequestReturnTaskBody": {
      "description": "MES请求物料返库任务信息实体类",
      "type": "object",
      "properties": {
        "TaskCode": {
          "description": "申请任务号",
          "type": "string"
        },
        "StationCode": {
          "description": "申请机台号",
          "type": "string"
        },
        "LineCode": {
          "description": "机台关联的输送线编号",
          "type": "string"
        },
        "ReqTime": {
          "description": "申请时间",
          "type": "string"
        },
        "MaterialList": {
          "description": "物料明细信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MesRequestMaterialBody"
          }
        }
      }
    },
    "MesStationRotationReportBody": {
      "description": "Mes发送机台旋转通知信息实体类",
      "type": "object",
      "properties": {
        "ReqTime": {
          "description": "申请时间",
          "type": "string"
        },
        "StationList": {
          "description": "机台信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/StationReportBody"
          }
        }
      }
    },
    "StationReportBody": {
      "description": "机台信息实体类",
      "type": "object",
      "properties": {
        "StationCode": {
          "description": "机台号",
          "type": "string"
        },
        "LineCode": {
          "description": "机台关联的输送线编号",
          "type": "string"
        }
      }
    },
    "MesRequestChangeInventoryBody": {
      "description": "Mes请求更改物料库存数量实体类",
      "type": "object",
      "properties": {
        "ReqTime": {
          "description": "申请时间",
          "type": "string"
        },
        "MaterialList": {
          "description": "申请信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ChangeInventoryBody"
          }
        }
      }
    },
    "ChangeInventoryBody": {
      "description": "要更改库存的明细信息",
      "type": "object",
      "properties": {
        "MaterialBarcode": {
          "description": "物料条码",
          "type": "string"
        },
        "Qty": {
          "format": "double",
          "description": "出库申请：申请数量；返库申请：WMS送料时传入的数量",
          "type": "number"
        }
      }
    },
    "ProductInboundRequestBody": {
      "description": "成品入库请求实体类",
      "type": "object",
      "properties": {
        "TaskCode": {
          "description": "申请任务号",
          "type": "string"
        },
        "StationCode": {
          "description": "机台号",
          "type": "string"
        },
        "MaterialPS": {
          "description": "包装类型",
          "type": "string"
        },
        "ReqTime": {
          "description": "申请时间",
          "type": "string"
        },
        "MaterialList": {
          "description": "入库成品信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductBody"
          }
        }
      }
    },
    "ProductBody": {
      "description": "入库成品信息",
      "type": "object",
      "properties": {
        "MaterialCode": {
          "description": "物料编码",
          "type": "string"
        },
        "MaterialBarcode": {
          "description": "箱条码",
          "type": "string"
        },
        "BatchNo": {
          "description": "批次号",
          "type": "string"
        },
        "Qty": {
          "description": "数量",
          "type": "string"
        }
      }
    },
...
  },
...
}